+++ src/__tests__/index.test.ts
@@ -123,6 +123,10 @@ describe('Emarsys core API', () => {
       expect(typeof Emarsys.geofence.disable).toBe('function');
       expect(typeof Emarsys.geofence.isEnabled).toBe('function');
       expect(typeof Emarsys.geofence.getRegisteredGeofences).toBe('function');
+
+      expect(typeof Emarsys.inbox.fetchMessages).toBe('function');
+      expect(typeof Emarsys.inbox.addTag).toBe('function');
+      expect(typeof Emarsys.inbox.removeTag).toBe('function');
     });
   });
 });