+++ src/index.ts
@@ -4,6 +4,7 @@ import { createConfigModule } from './config';
 import type { EmarsysApi, EmarsysPlugin } from './definitions';
 import { createGeofenceModule } from './geofence';
 import { createInAppModule } from './inApp';
+import { createInboxModule } from './inbox';
 import { createPushModule } from './push';
 
 const EmarsysPluginInstance = registerPlugin<EmarsysPlugin>('Emarsys');
@@ -19,7 +20,16 @@ const Emarsys: EmarsysApi = {
   inApp: createInAppModule(EmarsysPluginInstance),
   config: createConfigModule(EmarsysPluginInstance),
   geofence: createGeofenceModule(EmarsysPluginInstance),
+  inbox: createInboxModule(EmarsysPluginInstance),
 };
+
+// TODO: extract values from package.json
+void EmarsysPluginInstance.trackCustomEvent({
+  eventName: 'wrapper:init',
+  eventAttributes: { type: 'capacitor', version: '0.1.0', frameworkVersion: '8.5.0' },
+}).catch(() => {
+  /* ignore */
+});
 
 export * from './definitions';
 export { Emarsys };